Refactor to separate tab fxml (#9)
* started with work on refactoring

* refactoring working now

* added buttons

---------

+++ src/main/kotlin/com/nbk/weyay/weyaydesktopclient/Utils.kt
@@ -15,7 +15,7 @@ suspend fun checkReachability(destination: Destination): Destination {
     delay(duration.toLong())
     Socket().use {
         try {
-            it.connect(InetSocketAddress(destination.host, destination.port), TIMEOUT_MS)
+            it.connect(InetSocketAddress(destination.host, destination.port.toInt()), TIMEOUT_MS)
             destination.status = Status.REACHABLE
         } catch (e: IOException) {
             destination.status = Status.UNREACHABLE
